Address universal sandbox review findings
