+++ Tiltfile
@@ -121,6 +121,12 @@ preview_kro_context = 'kind-faros-kro'
 preview_kro_node = 'faros-kro-control-plane'
 universal_dev_image = 'ghcr.io/faroshq/faros-universal-dev:latest'
 universal_dev_image_repository = 'ghcr.io/faroshq/faros-universal-dev'
+# The local loop keeps the historical force default, while allowing an
+# explicit off/byo-only mode to avoid constructing any platform sandbox
+# resources. Keep this in Starlark as well as the shell fallback below so the
+# effective mode controls Tilt's resource graph (not only the provider env).
+app_studio_sandbox_mode = os.getenv('APP_STUDIO_RUN_SANDBOX_MODE', '').strip().lower() or 'force'
+app_studio_sandbox_force = app_studio_sandbox_mode == 'force'
 # Host address as seen FROM INSIDE the faros-kro containers. Resolve
 # host.docker.internal inside the node first: on Docker Desktop/OrbStack the
 # bridge gateway below is the VM, not the host, and dialing it gets connection
@@ -353,7 +359,10 @@ local_resource(
         'providers/app-studio/deploy/chart/values.yaml',
         'providers/app-studio/.env',
     ],
-    resource_deps=['hub', 'app-studio-db', 'dev-agent-image', 'universal-dev-image', 'app-studio-preview-console-key'],
+    resource_deps=(
+        ['hub', 'app-studio-db', 'dev-agent-image', 'app-studio-preview-console-key']
+        + (['universal-dev-image'] if app_studio_sandbox_force else [])
+    ),
     readiness_probe=probe(
         period_secs=5,
         http_get=http_get_action(port=8085, path='/healthz'),
@@ -498,9 +507,10 @@ local_resource(
 # Platform-curated Node/Go/Python toolchain image for the private coding
 # environment. It is separate from the injected dev-agent binary so the image
 # can be pinned independently in a production InfrastructureProvider CR.
-local_resource(
-    'universal-dev-image',
-    cmd=('''
+if app_studio_sandbox_force:
+    local_resource(
+        'universal-dev-image',
+        cmd=('''
 set -eu
 make load-universal-dev-image
 universal_digest="$(docker exec {kro_node} ctr -n k8s.io images ls 'name=={universal_image}' | awk 'NR == 2 {{ print $3 }}')"
@@ -514,17 +524,17 @@ esac
 docker exec {kro_node} ctr -n k8s.io images tag --force \
   {universal_image} {universal_repository}@$universal_digest
 ''').format(
-                   kro_node=preview_kro_node,
-                   universal_image=universal_dev_image,
-                   universal_repository=universal_dev_image_repository,
-               ),
-    deps=[
-        'providers/infrastructure/dev-agent/Dockerfile.universal',
-        'Makefile',
-    ],
-    resource_deps=['kro-mgmt-up'],
-    labels=['providers-kro'],
-)
+                       kro_node=preview_kro_node,
+                       universal_image=universal_dev_image,
+                       universal_repository=universal_dev_image_repository,
+                   ),
+        deps=[
+            'providers/infrastructure/dev-agent/Dockerfile.universal',
+            'Makefile',
+        ],
+        resource_deps=['kro-mgmt-up'],
+        labels=['providers-kro'],
+    )
 
 # --- providers-kuery (fleet query engine) ---
 # Local Postgres for the kuery store. Dev always runs the same SQL backend
@@ -703,6 +713,31 @@ local_resource(
     labels=['providers-kro'],
 )
 
+infrastructure_sandbox_digest_script = ''
+infrastructure_sandbox_env = 'FAROS_CODING_SANDBOX_ENABLED=false \\'
+infrastructure_resource_deps = [
+    'hub',
+    'kro-mgmt-up',
+    'preview-gateway-up',
+    'app-studio-preview-dns',
+    'app-studio-preview-port-forward',
+    'app-studio-preview-console-key',
+]
+if app_studio_sandbox_force:
+    infrastructure_sandbox_digest_script = ('''
+universal_digest="$(docker exec {kro_node} ctr -n k8s.io images ls 'name=={universal_image}' | awk 'NR == 2 {{ print $3 }}')"
+case "$universal_digest" in
+  sha256:????????????????????????????????????????????????????????????????) ;;
+  *) echo "local universal image has no immutable sha256 manifest digest" >&2; exit 1 ;;
+esac
+''').format(
+        kro_node=preview_kro_node,
+        universal_image=universal_dev_image,
+    )
+    infrastructure_sandbox_env = ('''FAROS_CODING_SANDBOX_ENABLED=true \\
+FAROS_DEV_IMAGE_UNIVERSAL="{universal_repository}@$universal_digest" \\''').format(universal_repository=universal_dev_image_repository)
+    infrastructure_resource_deps = infrastructure_resource_deps + ['universal-dev-image']
+
 local_resource(
     'infrastructure',
     cmd='make build-infrastructure-provider',
@@ -713,12 +748,7 @@ if [ -z "$host_gateway" ]; then
   host_gateway="$(docker inspect -f '{docker_gateway_format}' {kro_node})"
 fi
 test -n "$host_gateway"
-universal_digest="$(docker exec {kro_node} ctr -n k8s.io images ls 'name=={universal_image}' | awk 'NR == 2 {{ print $3 }}')"
-case "$universal_digest" in
-  sha256:????????????????????????????????????????????????????????????????) ;;
-  *) echo "local universal image has no immutable sha256 manifest digest" >&2; exit 1 ;;
-esac
-KRO_KUBECONFIG={kro_kubeconfig} \
+{sandbox_digest_script}KRO_KUBECONFIG={kro_kubeconfig} \
 FAROS_GATEWAY_NAME={gateway_name} \
 FAROS_GATEWAY_NAMESPACE={gateway_namespace} \
 FAROS_APP_BASE_DOMAIN={base_domain} \
@@ -726,20 +756,19 @@ FAROS_APP_PUBLIC_PORT={public_port} \
 FAROS_ACCESS_HUB_URL="https://$host_gateway:9443" \
 FAROS_ACCESS_HUB_PUBLIC_URL={hub_public_url} \
 FAROS_ACCESS_HUB_INSECURE=true \
-FAROS_CODING_SANDBOX_ENABLED=true \
-FAROS_DEV_IMAGE_UNIVERSAL="{universal_repository}@$universal_digest" \
+{sandbox_env}
 make run-provider-infrastructure
 ''').format(
                    docker_gateway_format=docker_gateway_format,
                    kro_node=preview_kro_node,
-                   universal_image=universal_dev_image,
                    kro_kubeconfig=preview_kro_kubeconfig,
                    gateway_name=preview_gateway_name,
                    gateway_namespace=preview_gateway_namespace,
                    base_domain=preview_app_base_domain,
                    public_port=preview_app_public_port,
                    hub_public_url=preview_hub_public_url,
-                   universal_repository=universal_dev_image_repository,
+                   sandbox_digest_script=infrastructure_sandbox_digest_script,
+                   sandbox_env=infrastructure_sandbox_env,
                ),
     deps=[
         'providers/infrastructure/main.go',
@@ -776,15 +805,7 @@ make run-provider-infrastructure
     # backend cluster the catalog reads from. Both must be green
     # before the provider starts; otherwise it boots in stub mode
     # which is fine but confusing for the dev who just ran `tilt up`.
-    resource_deps=[
-        'hub',
-        'kro-mgmt-up',
-        'preview-gateway-up',
-        'app-studio-preview-dns',
-        'app-studio-preview-port-forward',
-        'app-studio-preview-console-key',
-        'universal-dev-image',
-    ],
+    resource_deps=infrastructure_resource_deps,
     readiness_probe=probe(
         period_secs=5,
         http_get=http_get_action(port=8082, path='/healthz'),
